+++ build.gradle
@@ -38,7 +38,7 @@ project.ext.threeDotVersion = "3.3.0.1"
 project.ext.install4jDir = hasProperty("install4jDir") ? getProperty("install4jDir") : (OperatingSystem.current().isWindows() ? 'C:/Program Files/install4j6' : 'install4j6')
 sourceCompatibility = 1.8
 targetCompatibility = 1.8
-mainClassName = "net.sf.jabref.JabRefMain"
+mainClassName = "net.sf.jabref.JabRef"
 
 install4j {
     installDir = file(project.ext.install4jDir)
