+++ src/main/java/net/sf/jabref/gui/remote/JabRefMessageHandler.java
@@ -15,7 +15,7 @@
 */
 package net.sf.jabref.gui.remote;
 
-import net.sf.jabref.JabRef;
+import net.sf.jabref.JabRefGUI;
 import net.sf.jabref.cli.ArgumentProcessor;
 import net.sf.jabref.importer.ParserResult;
 import net.sf.jabref.logic.remote.server.MessageHandler;
@@ -26,7 +26,8 @@ public class JabRefMessageHandler implements MessageHandler {
 
     @Override
     public void handleMessage(String message) {
-        ArgumentProcessor argumentProcessor = new ArgumentProcessor(message.split("\n"), false);
+        ArgumentProcessor argumentProcessor = new ArgumentProcessor(message.split("\n"),
+                ArgumentProcessor.Mode.REMOTE_START);
         if (!(argumentProcessor.hasParserResults())) {
             throw new IllegalStateException("Could not start JabRef with arguments " + message);
         }
@@ -34,7 +35,7 @@ public class JabRefMessageHandler implements MessageHandler {
         List<ParserResult> loaded = argumentProcessor.getParserResults().get();
         for (int i = 0; i < loaded.size(); i++) {
             ParserResult pr = loaded.get(i);
-            JabRef.mainFrame.addParserResult(pr, i == 0);
+            JabRefGUI.mainFrame.addParserResult(pr, i == 0);
         }
     }
 }
