+++ src/main/java/net/sf/jabref/importer/EntryFromPDFCreator.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 import org.apache.pdfbox.pdmodel.PDDocument;
 import org.apache.pdfbox.pdmodel.PDDocumentInformation;
 
-import net.sf.jabref.JabRef;
+import net.sf.jabref.JabRefGUI;
 import net.sf.jabref.external.ExternalFileType;
 import net.sf.jabref.external.ExternalFileTypes;
 import net.sf.jabref.gui.IconTheme;
@@ -56,7 +56,7 @@ public class EntryFromPDFCreator extends EntryFromFileCreator {
             return Optional.empty();
         }
 
-        PdfImporter pi = new PdfImporter(JabRef.mainFrame, JabRef.mainFrame.getCurrentBasePanel(), JabRef.mainFrame.getCurrentBasePanel().mainTable, -1);
+        PdfImporter pi = new PdfImporter(JabRefGUI.mainFrame, JabRefGUI.mainFrame.getCurrentBasePanel(), JabRefGUI.mainFrame.getCurrentBasePanel().mainTable, -1);
         String[] fileNames = {pdfFile.toString()};
         ImportPdfFilesResult res = pi.importPdfFiles(fileNames);
         if (res.getEntries().size() == 1) {
