+++ src/main/java/net/sf/jabref/importer/fileformat/FreeCiteImporter.java
@@ -41,7 +41,7 @@ import net.sf.jabref.importer.OutputPrinter;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.model.entry.BibtexEntryTypes;
 import net.sf.jabref.Globals;
-import net.sf.jabref.JabRef;
+import net.sf.jabref.JabRefGUI;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.logic.labelpattern.LabelPatternUtil;
 import net.sf.jabref.model.entry.EntryType;
@@ -214,7 +214,7 @@ public class FreeCiteImporter extends ImportFormat {
                     e.setType(type);
 
                     // autogenerate label (BibTeX key)
-                    LabelPatternUtil.makeLabel(JabRef.mainFrame.getCurrentBasePanel().getBibDatabaseContext().getMetaData(), JabRef.mainFrame.getCurrentBasePanel().getDatabase(), e);
+                    LabelPatternUtil.makeLabel(JabRefGUI.mainFrame.getCurrentBasePanel().getBibDatabaseContext().getMetaData(), JabRefGUI.mainFrame.getCurrentBasePanel().getDatabase(), e);
 
                     res.add(e);
                 }
