+++ src/main/java/net/sf/jabref/specialfields/SpecialFieldDatabaseChangeListener.java
@@ -3,7 +3,7 @@ package net.sf.jabref.specialfields;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.model.database.DatabaseChangeEvent;
 import net.sf.jabref.model.database.DatabaseChangeListener;
-import net.sf.jabref.JabRef;
+import net.sf.jabref.JabRefGUI;
 import net.sf.jabref.gui.undo.NamedCompound;
 import net.sf.jabref.logic.l10n.Localization;
 
@@ -22,7 +22,7 @@ public class SpecialFieldDatabaseChangeListener implements
             NamedCompound nc = new NamedCompound(Localization.lang("Synchronized special fields based on keywords"));
             SpecialFieldsUtils.syncSpecialFieldsFromKeywords(entry, nc);
             nc.end();
-            JabRef.mainFrame.getCurrentBasePanel().undoManager.addEdit(nc);
+            JabRefGUI.mainFrame.getCurrentBasePanel().undoManager.addEdit(nc);
         }
     }
