+++ src/main/java/net/sf/jabref/util/ResourceExtractor.java
@@ -20,11 +20,11 @@ import java.io.File;
 import java.io.IOException;
 import java.net.URL;
 
+import net.sf.jabref.JabRef;
 import net.sf.jabref.gui.net.MonitoredURLDownload;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 
-import net.sf.jabref.JabRef;
 import net.sf.jabref.gui.worker.Worker;
 import net.sf.jabref.logic.net.URLDownload;
